Move all blackbox tests to a dedidated package, where they will not be tempted to use package private things.
