+++ src/test/java/blackbox/simulations/TimeSpreadExpirationSim.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stormpot.simulations;
+package blackbox.simulations;
 
 import stormpot.*;
