+++ src/test/java/blackbox/slow/PoolIT.java
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package stormpot.slow;
+package blackbox.slow;
 
+import extensions.ExecutorExtension;
+import extensions.FailurePrinterExtension;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
