+++ src/test/java/stormpot/ReallocatingAdaptorTest.java
@@ -15,6 +15,7 @@
  */
 package stormpot;
 
+import extensions.FailurePrinterExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
